인증 / 인가 추가 - 회원가입시 비밀번호 암호화 적용, JWT를 통한 게시물, 댓글 수정권한 부여, 로그인시 JWT 발급

+++ src/main/kotlin/com/sparta/wangnyang/domain/board/service/BoardServiceImpl.kt
@@ -45,9 +45,12 @@ class BoardServiceImpl (
 
     // 게시물 수정
     @Transactional
-    override fun updateBoard(boardId: Long, request: UpdateBoardRequest): BoardResponse {
+    override fun updateBoard(userId:String, boardId: Long, request: UpdateBoardRequest): BoardResponse {
         val board = boardRepository.findByIdOrNull(boardId) ?: throw Exception("게시물이 존재하지 않습니다.")
         val (title,mainText,writer) = request
+
+        //본인 게시글이 아니면 에러를 발생시킨다.
+        if(userId != board.userId && writer != board.userId) throw Exception("해당 게시물의 작성자가 아닙니다.")
 
         board.title = title
         board.mainText = mainText
@@ -59,8 +62,11 @@ class BoardServiceImpl (
 
     // 게시뭏 삭제
     @Transactional
-    override fun deleteBoard(boardId: Long) {
+    override fun deleteBoard(userId: String, boardId: Long) {
         val board = boardRepository.findByIdOrNull(boardId) ?: throw Exception("게시물이 존재하지 않습니다.")
+
+        if(board.userId != userId) throw Exception("해당 게시물의 작성자가 아닙니다.")
+
         boardRepository.delete(board)
     }
